+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 3.2.12 (Unpublished)
 ### Bugfix
  - hide `shepherdJS` when it is not needed
+ - #22 - `surveyJS` check if data is saved before going to the next page. If not saved show an error message.
 
 # 3.2.11
 ### Bugfix